requirements.txt
python-dateutil==2.5.3
folium==0.2.1
numpy==1.11.1
PyOpenGL==3.1.0
PyQt5==5.7
